requirements.txt
Babel==2.7.0
Django==2.2.7
django-ajax-selects==1.7.1
django-autocomplete-light==3.4.1
django-bootstrap3==11.1.0
django-crispy-forms==1.8.0
django-phonenumber-field==3.0.1
django-popup-field==0.1.6
django-tempus-dominus==5.1.2.9
phonenumbers==8.10.23
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
